requirements.txt
asgiref==3.2.5
certifi==2020.4.5.1
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.2
pytz==2019.3
sqlparse==0.3.1
whitenoise==4.1.4
